"Changed parameters command to allow for multiple fits in a single pysavuka session. Added a clear command to clear commandline of text. Added a buggy chi_error command. Changed fit routine to only expect a 2D array for data argument. Added docstrings to all functions in fit.py. Added headless parameter creation in params.py to allow for fits to be run multiple times without prompting user for parameters each time. Added FIG_NUMBER constant to plot_funcs.py to guarantee unique plot numbers for each plotted window. Added general plot_xy function to plot arbitrary numpy arrays against each other. Got rid of savuka buffer 'hashing'. Added tests for params."

+++ src/buffer.py
@@ -18,11 +18,11 @@ class Dimension(object):
         return "(" + self.name + ", " + str(self.data) + ")\n"
 
     def __init__(self, data=None, name=None):
-        self.data = np.asarray(data) if data is not None else np.asarray([])
+        self.data = np.asarray(data, dtype=np.float64) if data is not None else np.asarray([])
         self.name = name if name is not None else ''
 
     def set_data(self, new_data):
-        self.data = np.asarray(new_data)
+        self.data = np.asarray(new_data, dtype=np.float64)
 
     def set_name(self, new_name):
         self.name = str(new_name)
